fix(rubocop): extract search instrumentation helper modules
Second slice of warehouse 9af971264: move Core, PayloadHelpers,
ClassicResultSummaries, and ResultSummaries into nested modules under
Search::Instrumentation. Thin shell keeps constants and extends all
event/helper modules. AiUsage remains on ApiEvents from slice 1.

Issue: No ticket/issue
